+++ src/main.rs
@@ -10,7 +10,7 @@ use scanner::start_scan;
 fn main() {
     let args = Args::parse();
 
-    let results = start_scan(args.from, args.to, args.port, args.threads as u32, None);
+    let results = start_scan(args.from, args.to, args.port, args.threads, None);
 
     for result in results {
         println!("{:?}", result);
